Part 1 done

+++ src/day9/mod.rs
@@ -1,11 +1,11 @@
 use itertools::Itertools;
 
-pub fn find_invalid_number(input: Vec<i32>, preamble: usize) -> i32
+pub fn find_invalid_number(input: Vec<u64>, preamble: usize) -> u64
 {
   let mut start = preamble;
 
   while start < input.len() {
-    let preamble_sums : Vec<i32> = input[(start-preamble)..start].iter().map(|x| *x).collect::<Vec<i32>>().iter().combinations(2).map(|x| x[0] + x[1]).collect();
+    let preamble_sums : Vec<u64> = input[(start-preamble)..start].iter().map(|x| *x).collect::<Vec<u64>>().iter().combinations(2).map(|x| x[0] + x[1]).collect();
 
     if !preamble_sums.contains(&input[start]) {
       return input[start];
